get started on property step
