+++ step/math/MinStep.h
@@ -3,6 +3,7 @@
 #define MIN_STEP 0x60
 
 #include "step/TraversalStep.h"
+#include "step/modulate/ReductionStep.h"
 #include "traversal/Traverser.h"
 #include "traversal/Scope.h"
 #include "traversal/Comparison.h"
@@ -11,22 +12,17 @@
 
 namespace gremlinxx {
 
-    class MinStep : public TraversalStep {
-        private:
-            std::optional<ScopeContext> scope_context;
-
-            void apply_global(GraphTraversal* traversal, gremlinxx::traversal::TraverserSet& traversers);
-
-            void apply_local(GraphTraversal* traversal, gremlinxx::traversal::TraverserSet& traversers);
-
+    class MinStep : public ReductionStep {
         public:
             MinStep();
 
-            inline void set_scope_context(ScopeContext sc) { this->scope_context = sc; }
+            using ReductionStep::apply_global;
+            virtual void apply_global(GraphTraversal* traversal, gremlinxx::traversal::TraverserSet& traversers);
 
-            virtual void apply(GraphTraversal* traversal, gremlinxx::traversal::TraverserSet& traversers);
+            using ReductionStep::apply_local;
+            virtual void apply_local(GraphTraversal* traversal, gremlinxx::traversal::TraverserSet& traversers);
 
-            using TraversalStep::getInfo;
+            using ReductionStep::getInfo;
             virtual std::string getInfo();
     };
